Added code to pad out runs that end early.
Successful runs are terminated when they find the desired solution, so many of the runs (especially for lexicase) have fewer than 300 generations. These versions of the notebooks look at the possibility of padding out runs that end early by either replicating the data (including the number of clusters and the diversity), or padding with generations having 1 cluster and 0 diversity.